implement clustering protocol; add a base protocol
